Remove assumption about the roles path
This removes the assumption that roles for a playbook are installed in a
directory called `roles` beneath `playbook.yml` (which can make the including
of os-specific or version-specific variables fail).

Instead, this assumes that all roles are checked out in a same directory.
This will break implementations where roles that use silpion.lib are installed
in a separate directory, but this use case should be less common than having
oles installed somewhere else than a directory called `roles` beneath
`playbook.yml`.